Function sorting::simplesorts::bubble::sort

pub fn sort<T: PartialOrd>(input: &mut [T])

Simple sort: bubble sort.

Details

Bubble sort will "bubble" elements from a lower poisition to their sorted position. It is a bad sorting algorithm but extremely simple to implement.

Scaling

Two nested loops are the basis of bubble sort. The first one indicates the number of elements to skip from the end in the second loop. The second loop iterates over elements, starting at the begining of the vector and ending at the index of the outer loop.

Because of these two loops, the vector is completely visited for every elements, hence the algorithm's O(N2) scaling.

If the vector is already sorted, no swap will take place and this can be detected at the end of the first inner loop. As such, the best case scenario for bubble sort is O(N) when the vector is already sorted.

Optimizations

Detect if no swap occurred during the inner loop. In that case, break the outer loop. This makes the function returns as soon as the vector is sorted.

If the vector is initially sorted, the best case scenario of O(N) scaling is then achieved.

Notes

The type T of the vector elements to sort must implement the PartialOrd trait so the compiler knows how to compare the elements and sort them.

Just don't use this sorting function. It's purely academic.

Examples

let mut data: Vec<i32> = vec![4, 2, 3, 1, 5];
sorting::simplesorts::bubble::sort(&mut data);
assert_eq!(vec![1, 2, 3, 4, 5], data);